ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
pdfminer.six==20221105
scikit-learn==1.3.2
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
